Fix code issues identified by Oracle Dev. Studio
(+ a couple identified by GCC 6)

Unreachable statements, variable scope hiding, etc.

+++ server/vglconfigLauncher.h
@@ -1,5 +1,5 @@
 /* Copyright (C)2007 Sun Microsystems, Inc.
- * Copyright (C)2014 D. R. Commander
+ * Copyright (C)2014, 2018 D. R. Commander
  *
  * This library is free software and may be redistributed and/or modified under
  * the terms of the wxWindows Library License, Version 3.1 or (at your option)
@@ -58,7 +58,7 @@ namespace vglserver
 					unsetenv("LD_PRELOAD_32");
 					unsetenv("LD_PRELOAD_64");
 					sprintf(commandLine, "%s -display %s -shmid %d -ppid %d",
-						fconfig.config, DisplayString(dpy), shmid, getpid());
+						fconfig.config, DisplayString(dpy), shmid, (int)getpid());
 					if(system(commandLine) == -1) _throwunix();
 				}
 				catch(vglutil::Error &e)
